+++ src/App.tsx
@@ -1,13 +1,14 @@
 import { GoogleOAuthProvider } from "@react-oauth/google";
 import "./App.css";
-import "./App.css";
 
 const clientId = import.meta.env.VITE_GOOGLE_CLIENT_ID;
 
 function App() {
-  <GoogleOAuthProvider clientId={clientId}>
-    return <></>;
-  </GoogleOAuthProvider>;
+  return (
+    <GoogleOAuthProvider clientId={clientId}>
+      <></>
+    </GoogleOAuthProvider>
+  );
 }
 
 export default App;
