+++ src/Home.tsx
@@ -1,5 +1,3 @@
-const Home = () => {
+export default function Home() {
   return <div>Home</div>;
-};
-
-export default Home;
+}
